+++ app/src/main/java/com/benefit/ui/fragments/ProductFragment.java
@@ -1,6 +1,4 @@
-package com.benefit.ui;
-
-import androidx.lifecycle.ViewModelProviders;
+package com.benefit.ui.fragments;
 
 import android.os.Bundle;
 
@@ -13,28 +11,32 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.benefit.R;
+import com.benefit.model.Product;
 import com.benefit.model.User;
 
-public class ProfileFragment extends Fragment {
+public class ProductFragment extends Fragment {
 
 
-    public static ProfileFragment newInstance(User user) {
-        ProfileFragment fragment = new ProfileFragment();
+    public static ProductFragment newInstance(Product product, User user) {
+        ProductFragment fragment = new ProductFragment();
+        fragment.product = product;
         fragment.user = user;
-        return new ProfileFragment();
+        return fragment;
     }
 
+    private Product product;
     private User user;
 
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.profile_fragment, container, false);
+        return inflater.inflate(R.layout.product_fragment, container, false);
     }
 
     @Override
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
+
     }
 
 }
