+++ app/src/main/res/layout/activity_sign_in.xml
@@ -169,7 +169,7 @@
 
             <fragment
                 android:id="@+id/googleMapFragment"
-                class="com.benefit.ui.WorkaroundMapFragment"
+                class="com.benefit.ui.fragments.WorkaroundMapFragment"
                 android:layout_width="match_parent"
                 android:layout_height="250dp"
                 android:layout_gravity="center_horizontal"
