+++ app/src/main/res/layout/all_products_fragment.xml
@@ -7,7 +7,7 @@
     android:background="@android:color/white"
     android:layoutDirection="ltr"
     android:orientation="vertical"
-    tools:context=".ui.AllProductsFragment">
+    tools:context=".ui.fragments.AllProductsFragment">
 
     <androidx.appcompat.widget.SearchView
         android:id="@+id/search_input"
