+++ app/src/main/res/layout/give_fragment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.GiveFragment">
+    tools:context=".ui.fragments.GiveFragment">
 
     <ScrollView
         android:id="@+id/scroll_give"
