+++ app/src/main/res/layout/product_fragment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.ProductFragment">
+    tools:context=".ui.fragments.ProductFragment">
 
     <TableLayout
         android:id="@+id/product_page_back_area"
