+++ app/src/main/res/layout/profile_fragment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.ProfileFragment">
+    tools:context=".ui.fragments.ProfileFragment">
 
     <LinearLayout
         android:id="@+id/greeting_and_compliment_layout"
